refactor proposal version command routes
