+++ src/app/router_registry.py
@@ -171,6 +171,7 @@ from app.routers.proposal_memo_reporting import router as proposal_memo_reportin
 from app.routers.proposal_memos import router as proposal_memos_router
 from app.routers.proposal_narratives import router as proposal_narratives_router
 from app.routers.proposal_operations import router as proposal_operations_router
+from app.routers.proposal_version_commands import router as proposal_version_commands_router
 from app.routers.proposal_versions import router as proposal_versions_router
 from app.routers.proposal_workflow import router as proposal_workflow_router
 from app.routers.proposal_workflow_decisions import router as proposal_workflow_decisions_router
@@ -244,6 +245,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(proposals_router)
     app.include_router(proposal_operations_router)
     app.include_router(proposal_versions_router)
+    app.include_router(proposal_version_commands_router)
     app.include_router(proposal_workflow_router)
     app.include_router(proposal_workflow_decisions_router)
     app.include_router(proposal_workflow_evidence_router)
