+++ src/app/routers/proposal_versions.py
@@ -1,10 +1,7 @@
-from fastapi import APIRouter, Header, Path, Query
+from fastapi import APIRouter, Path, Query
 
 from app.contracts.proposals import (
-    ProposalBodyRequest,
-    ProposalCreateEnvelopeResponse,
     ProposalEnvelopeResponse,
-    ProposalVersionCreateRequest,
     ProposalVersionEnvelopeResponse,
 )
 from app.middleware.correlation import correlation_id_var
@@ -73,66 +70,5 @@ async def get_proposal_version_replay_evidence(
     return await service.get_proposal_version_replay_evidence(
         proposal_id=proposal_id,
         version_no=version_no,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/versions",
-    response_model=ProposalCreateEnvelopeResponse,
-    summary="Create Proposal Version",
-    description="Creates the next persisted version for an existing advisory proposal.",
-)
-async def create_proposal_version(
-    request: ProposalVersionCreateRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal-version creation requests.",
-        examples=["idem-version-2"],
-    ),
-) -> ProposalCreateEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.create_proposal_version(
-        proposal_id=proposal_id,
-        body=request.body,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/versions/async",
-    response_model=ProposalEnvelopeResponse,
-    summary="Create Proposal Version Asynchronously",
-    description=(
-        "Starts an asynchronous version-create operation in lotus-advise for an existing "
-        "proposal. Gateway returns source-owned operation posture only."
-    ),
-)
-async def create_proposal_version_async(
-    request: ProposalBodyRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for async proposal-version creation.",
-        examples=["idem-version-async-1"],
-    ),
-) -> ProposalEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.create_proposal_version_async(
-        proposal_id=proposal_id,
-        body=request.body,
-        idempotency_key=idempotency_key,
         correlation_id=correlation_id,
     )
